+++ python/.gitignore
@@ -1,0 +1,1 @@
+doc/_build/
