+++ python/doc/_templates/globaltoc.html
@@ -4,7 +4,8 @@
 <ul>
   <li><a href="{{ pathto('index') }}">Home</a></li>
   <li><a href="{{ pathto('install') }}">Install</a></li>
-  <li><a href="{{ pathto('api') }}">API Documentation</a></li>
+  <li><a href="{{ pathto('api') }}">High level API</a></li>
+  <li><a href="{{ pathto('capi') }}">Low level API</a></li>
 </ul>
 <b>External Links:</b>
 <ul>
