+++ python/doc/_templates/sidebarintro.html
@@ -1,4 +1,4 @@
 <h3>deltachat {{release}}</h3>
 <p>
-  Bindings for <a href="https://delta.chat">Delta.chat</a>
+<a href="https://delta.chat">Delta.chat</a> Python Bindings
 </p>
