+++ python/doc/api.rst
@@ -1,14 +1,26 @@
 
-deltachat API Reference
-==============================
+deltachat high level API Reference
+===================================
+
+.. note::
+
+    This API is work in progress and may change in versions prior to 1.0.
 
 .. autosummary::
 
    deltachat.account
+   deltachat.chatting
 
 account module
 --------------
 
 .. automodule:: deltachat.account
+    :members:
+
+
+chatting module
+---------------
+
+.. automodule:: deltachat.chatting
     :members:
