+++ python/doc/capi.rst
@@ -1,0 +1,7 @@
+
+Low Level API Reference
+===================================
+
+.. automodule:: deltachat.capi.lib
+    :members:
+
