+++ python/src/deltachat/__init__.py
@@ -2,7 +2,6 @@ from deltachat import capi
 from deltachat.capi import ffi
 from deltachat.account import Account  # noqa
 
-
 __version__ = "0.5.dev0"
